Remove no-op exports
Any change to the environment that happens in the context of `system`
only affects the process spawned by `system`, not the main ruby process.

The only way to alter the environment of the ruby process is to modify
the `ENV` hash. It's unclear if that is necessary, since the existing
code wasn't actually doing anything.